Hitbox Manipulation Groundwork
Added some stuff in the Quinn and Manny animation blueprints for hitbox manipulation. Added some hitboxes to the base character themselves to get ready for the combo system. With the current implementation, an animation can trigger an event that will move the hitboxes to appropriate positions through the moveset data table. Right now, the event is on the Running animation for troubleshooting but can be called from any animation with an animation notify. One thing to note is that this has some issues with blendspaces, where it doesn't work even though as far as I understand it should still call the event. Additionally, some weirdness will happen when the animation starts after the trigger. This should only be a minor problem and will be gone once more finalized animations and controls are implemented.